Splitting centerPanelPages component
