+++ webapp/src/components/centerPanelPages/breadcrumbs.scss
@@ -1,0 +1,6 @@
+.Breadcrumbs {
+    flex-grow: 1;
+    .page-breadcrumb {
+        cursor: pointer;
+    }
+}
